Extracted error, api modules
